// frontend/src/lib/api/_generated/orchestrator.d.ts
export interface components {
    schemas: {
// ...
        FixedBondResult: {
            /** Npv */
            npv?: number | null;
            /** Clean Price */
            clean_price?: number | null;
            /** Dirty Price */
            dirty_price?: number | null;
            /** Accrued */
            accrued?: number | null;
            /** Yield To Maturity */
            yield_to_maturity?: number | null;
            /** Cashflows */
            cashflows?: {
                [key: string]: unknown;
            }[];
            /** Extras */
            extras?: {
                [key: string]: unknown;
            };
        };
// ...
        FloatingBondResult: {
            /** Npv */
            npv?: number | null;
            /** Clean Price */
            clean_price?: number | null;
            /** Dirty Price */
            dirty_price?: number | null;
            /** Accrued */
            accrued?: number | null;
            /** Cashflows */
            cashflows?: {
                [key: string]: unknown;
            }[];
            /** Extras */
            extras?: {
                [key: string]: unknown;
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}